Reorganize submodule popup, update functions
It might be better to set the update method by option in the popup
rather than prompting each time, but on the other hand, I'm not really
sure what the best default is.  Git's default is checkout which isn't so
great since it leaves the submodule in a detached state.  Git doesn't
support a global submodule.update variable, only a per-module
submodule.<name>.update setting.